Clear Slabsets before loading config (#2805)
